Navigate to home page after login

+++ src/components/AuthForm/SignInForm.tsx
@@ -1,11 +1,21 @@
 import React, {useState} from "react";
 import "./AuthForm.css";
-import {Alert, FormControl, FormHelperText, IconButton, InputAdornment, OutlinedInput} from "@mui/material";
+import {
+    Alert,
+    Checkbox,
+    FormControl,
+    FormControlLabel,
+    FormHelperText,
+    IconButton,
+    InputAdornment,
+    OutlinedInput,
+} from "@mui/material";
 import {validateEmail, validatePassword} from "./AuthValidator.ts";
 import {Visibility, VisibilityOff} from "@mui/icons-material";
 import {signIn} from "../../api/Auth.ts";
 import {LoadingButton} from "@mui/lab";
 import LoginIcon from "@mui/icons-material/Login";
+import useAppNavigator from "../../hooks/UseAppNavigator.ts";
 
 export default function SignInForm() {
     const [email, setEmail] = useState<string>("");
@@ -34,8 +44,20 @@ export default function SignInForm() {
     };
     const isPasswordValid = password !== "" && passwordError === null;
 
+    const [keepSignedIn, setKeepSignedIn] = useState<boolean>(false);
+    const handleKeepSignedInChange = () => {
+        setKeepSignedIn((prev) => !prev);
+    };
+    const keepSignedInButtonStyle = {
+        "& .MuiFormControlLabel-label": {
+            color: "#555",
+            fontFamily: "\"Trirong\", serif",
+        },
+    };
+
     const [isLoading, setIsLoading] = useState<boolean>(false);
     const [isSuccess, setIsSuccess] = useState<boolean | null>(null);
+    const {goToHome} = useAppNavigator();
     const canSubmit = isEmailValid && isPasswordValid;
     const handleSubmit = async (e: React.FormEvent<HTMLFormElement>) => {
         e.preventDefault();
@@ -43,8 +65,15 @@ export default function SignInForm() {
         setIsLoading(true);
         await signIn({email, password})
             .then((data) => {
-                localStorage.setItem("authToken", JSON.stringify(data.data.token));
+                if (keepSignedIn) {
+                    localStorage.setItem("authToken", JSON.stringify(data.data.token));
+                } else {
+                    sessionStorage.setItem("authToken", JSON.stringify(data.data.token));
+                }
+
                 setIsSuccess(true);
+
+                goToHome();
             })
             .catch((error) => {
                 console.log(error);
@@ -55,17 +84,12 @@ export default function SignInForm() {
     };
 
     const signInAlert = () => {
-        if (isSuccess === null) {
-            return null;
-        }
-
-        if (!isSuccess) {
+        if (isSuccess !== null && !isSuccess) {
             return <Alert severity="error">Invalid email or password. Please try again.</Alert>;
         }
 
-        return <Alert severity="success">Sign in successfully</Alert>;
+        return null;
     };
-
 
     return (
         <div className="auth-container">
@@ -117,6 +141,11 @@ export default function SignInForm() {
                         {!!passwordError && <FormHelperText error>{passwordError}</FormHelperText>}
                     </FormControl>
                 </div>
+
+                <FormControlLabel sx={keepSignedInButtonStyle}
+                                  label="Keep me signed in"
+                                  control={<Checkbox checked={keepSignedIn} onChange={handleKeepSignedInChange}/>}
+                />
 
                 {signInAlert()}
